atualizando os tipos de Intenger para Long nas operações envolvendo alunos

+++ src/main/java/com/br/SAM_FullStack/SAM_FullStack/dto/AdicionarAlunoDTO.java
@@ -6,8 +6,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AdicionarAlunoDTO {
-        private Integer idAdmin;
+        private Long idAdmin;
         private long idGrupo;
-        private Integer idAluno;
+        private Long idAluno;
 }
